Update to work with mostly the default KiCAD items or items downloaded via the KiCAD Libraries too.
2 Resistors added of 1k on the lines for TRUE and FALSE pull ups.

Several things deleted which appeared at Graphic shapes on the PCB sometimes
but could not be edited. this was done manually at the bottom of the PCB_KiCAD file.

Deleted some PCB items around the order which marked the Copper limits of top and bottom. may have been tied to TRUE & FALSE.

Changed USB socket type.

This has been done not to change the board but to aid understanding and allow anyone who wants slightly easier access.

I have not tested it and provide no warranty.

Signed-off-by: John Bradley <flypie@rocketmail.com>